Great Sand Dunes National Park, National Trails Day
A must visit! We had a wonderful time hiking the Mosca Pass trail and admiring
the dunes from afar before jumping in and getting _super sandy_. It might be a
little more crowded but going when the creek is flowing is pretty incredible.
Excellent way to spend National Trails Day.